fixes add be and add factDeliveredAt to orders table

+++ BuildingWorks.Infrastructure/Entities/Joininig/OrderMaterial.cs
@@ -4,7 +4,7 @@ namespace BuildingWorks.Infrastructure.Entities.Joininig;
 
 public class OrderMaterial
 {
-    public float PricePerOne { get; set; }
+    public decimal PricePerOne { get; set; }
     public int Quantity { get; set; }
 
     public Guid OrdersId { get; set; }
